+++ react/src/routeTree.gen.ts
@@ -12,10 +12,10 @@
 
 import { Route as rootRoute } from './routes/__root'
 import { Route as ResetPasswordImport } from './routes/reset-password'
-import { Route as RecoverPasswordImport } from './routes/recover-password'
 import { Route as LoginImport } from './routes/login'
 import { Route as LayoutImport } from './routes/_layout'
 import { Route as LayoutIndexImport } from './routes/_layout/index'
+import { Route as ResetPasswordTokenImport } from './routes/reset-password_/$token'
 import { Route as RegisterTokenImport } from './routes/register/$token'
 import { Route as LayoutSettingsImport } from './routes/_layout/settings'
 import { Route as LayoutGroupsImport } from './routes/_layout/groups'
@@ -27,11 +27,6 @@ import { Route as LayoutGroupsGroupIdLoopsImport } from './routes/_layout/groups
 
 const ResetPasswordRoute = ResetPasswordImport.update({
   path: '/reset-password',
-  getParentRoute: () => rootRoute,
-} as any)
-
-const RecoverPasswordRoute = RecoverPasswordImport.update({
-  path: '/recover-password',
   getParentRoute: () => rootRoute,
 } as any)
 
@@ -48,6 +43,11 @@ const LayoutRoute = LayoutImport.update({
 const LayoutIndexRoute = LayoutIndexImport.update({
   path: '/',
   getParentRoute: () => LayoutRoute,
+} as any)
+
+const ResetPasswordTokenRoute = ResetPasswordTokenImport.update({
+  path: '/reset-password/$token',
+  getParentRoute: () => rootRoute,
 } as any)
 
 const RegisterTokenRoute = RegisterTokenImport.update({
@@ -92,10 +92,6 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof LoginImport
       parentRoute: typeof rootRoute
     }
-    '/recover-password': {
-      preLoaderRoute: typeof RecoverPasswordImport
-      parentRoute: typeof rootRoute
-    }
     '/reset-password': {
       preLoaderRoute: typeof ResetPasswordImport
       parentRoute: typeof rootRoute
@@ -114,6 +110,10 @@ declare module '@tanstack/react-router' {
     }
     '/register/$token': {
       preLoaderRoute: typeof RegisterTokenImport
+      parentRoute: typeof rootRoute
+    }
+    '/reset-password/$token': {
+      preLoaderRoute: typeof ResetPasswordTokenImport
       parentRoute: typeof rootRoute
     }
     '/_layout/': {
@@ -143,9 +143,9 @@ export const routeTree = rootRoute.addChildren([
     LayoutGroupsGroupIdLoopsRoute,
   ]),
   LoginRoute,
-  RecoverPasswordRoute,
   ResetPasswordRoute,
   RegisterTokenRoute,
+  ResetPasswordTokenRoute,
 ])
 
 /* prettier-ignore-end */
